Fix 34 critical and high-severity bugs across the codebase
Critical fixes (5):
- NativeIndex: tombstones now persist through SSTable flush using \0 sentinel
- HTTP metadata: fix dangling references in invoke_on lambda captures
- Index WAL: fix CRC32/CRC32C polynomial mismatch in software fallback
- HTTP server: call std::_Exit on shutdown timeout to prevent use-after-free
- Auth: wire checkAuth middleware into all HTTP handler registerRoutes

High-severity fixes (29):
- TSM: fix integer sparse index offsets (52/60 → 56/64) for LATEST/FIRST
- TSM: guard readSparseIndex against files < 13 bytes
- TSM: capture tlStringDict before co_await to prevent stale dictionary
- TSM: validate header magic "TASM" on open
- ALP: validate num_blocks fits in 16-bit header field
- ALP: bounds-check exception_count against block_count
- HTTP write: guard against empty timestamps vector OOB access
- SIMD: use HWY_MAX_LANES_D for ComputeHistogram bins (ARM SVE safety)
- Compactor: return CompactionResult directly (eliminate stats data race)
- Compactor: use when_all_succeed instead of when_all
- Compactor: add RAII temp file cleanup on failure
- String encoder: copy thread-local buffer instead of move
- Manifest: save fileNumber before std::move to prevent use-after-move
- Manifest: atomic addFile+removeFiles in single fsync
- Functions: escape JSON in monitoring metrics and error responses
- Functions: fix createErrorResponse returning nullptr
- AlignedBuffer: add overflow checks to grow_uninit and write_bytes
- Tombstone: fix fd leak using exception_ptr cleanup pattern
- Tombstone: overflow-safe adjacency check for UINT64_MAX endTime
- Interpolation: cast to double before uint64_t subtraction in extrapolate
- FunctionUtils: fix interpolateLinear O(n*m) by hoisting search index
- TSM writer: accumulate int64 block sum as double to avoid signed overflow
- SSTable: verify decompressed size matches expected after ZSTD
- SSTable: cap uncompressedSize allocation at 64MB
- BoolEncoderRLE: throw on exhausted varint instead of returning 0
- Engine: add metrics increment to insertBatch
- Engine: pass seriesId through to query runner
- WAL file manager: guard against double WAL removal with file_exists
- TSM file manager: write to .tmp then atomic rename
- Line parser: own tag strings instead of dangling string_view

123 regression tests added across 23 new test files.
All 2981 tests pass.

+++ lib/functions/interpolation_functions.cpp
@@ -146,8 +146,10 @@ seastar::future<FunctionResult<double>> LinearInterpolationFunction::execute(con
                 if (input.count >= 2) {
                     double slope = (input.valueAt(1) - input.valueAt(0)) /
                                    static_cast<double>(input.timestampAt(1) - input.timestampAt(0));
+                    // Bug #22 fix: Cast both operands to double before subtraction
+                    // to prevent unsigned underflow when targetTime < timestampAt(0)
                     double extrapolated =
-                        input.valueAt(0) + slope * static_cast<double>(targetTime - input.timestampAt(0));
+                        input.valueAt(0) + slope * (static_cast<double>(targetTime) - static_cast<double>(input.timestampAt(0)));
                     result.values.push_back(extrapolated);
                 } else {
                     result.values.push_back(input.valueAt(0));
